refactor(parquet): adjust run_rows default value and implement background memory management for Linux
